asset-api/requirements.txt
fastapi 
uvicorn 
pycryptodome
PyJWT
python-dotenv
python-jose
passlib
sqlalchemy